PaginationVisuals - Utilise Class Inheritance for Tests
